charge_manager: Fix allocation_interval as micros.

+++ software/src/modules/charge_manager/current_allocator.cpp
@@ -1453,7 +1453,7 @@ int allocate_current(
                 charger.allocated_energy_this_rotation = 0;
             } else {
                 auto amps = (float)current_to_set / 1000.0f * phases_to_set;
-                auto amp_hours = amps * (float)(int64_t)(cfg->allocation_interval / 1_h);
+                auto amp_hours = amps * ((float)(int64_t)cfg->allocation_interval) / ((float)(int64_t)1_h);
                 auto watt_hours = amp_hours * 230.0f;
                 auto allocated_energy = watt_hours / 1000;
                 charger.allocated_energy_this_rotation += allocated_energy;
